Remove structs that are already in fmlib
