+++ pyropod/ropod/structs/__init__.py
@@ -1,4 +1,2 @@
 __title__ = 'structs'
 __version__ = '1.0.0'
-
-from .area import Area, SubArea
