+++ pyropod/ropod/utils/models.py
@@ -1,7 +1,5 @@
 from ropod.utils.timestamp import TimeStamp
 from ropod.utils.uuid import generate_uuid
-from ropod.structs.task import Task, TaskRequest
-from ropod.structs.elevator import ElevatorRequest, RobotCallUpdate, RobotElevatorCallReply
 
 meta_model_template = 'ropod-%s-schema.json'
 
@@ -72,53 +70,3 @@ class MessageFactoryBase(object):
         payload = {'payload': {'receivedMsg': message_id}}
         msg.update(payload)
         return msg
-
-
-class RopodMessageFactory(MessageFactoryBase):
-    def __init__(self):
-        super().__init__()
-
-        self.register_msg(Task.__name__, self)
-        self.register_msg(TaskRequest.__name__, self)
-        self.register_msg(ElevatorRequest.__name__, self)
-        self.register_msg(RobotCallUpdate.__name__, self)
-        self.register_msg(RobotElevatorCallReply.__name__, self)
-
-    def create_message(self, contents, recipients=[]):
-        if isinstance(contents, Task):
-            model = 'TASK'
-        elif isinstance(contents, TaskRequest):
-            model = 'TASK-REQUEST'
-        elif isinstance(contents, ElevatorRequest):
-            model = 'ELEVATOR-CMD'
-        elif isinstance(contents, RobotCallUpdate):
-            model = 'ROBOT-CALL-UPDATE'
-        elif isinstance(contents, RobotElevatorCallReply):
-            model = 'ROBOT-ELEVATOR-CALL-REPLY'
-
-        msg = self.get_header(model, recipients=recipients)
-        payload = self.get_payload(contents, model.lower())
-        msg.update(payload)
-        return msg
-
-    def get_query_msg(self, msg_type, payload_key, payload_value, success, receiverId):
-        '''Returns a dictionary representing a query response for the given
-        message type.
-        Assumption: header uses 'ropod-msg-schema.json' meta model
-
-        Keyword arguments:
-        :msg_type: a string representing a message type
-        :payload_key: string representing key for payload
-        :payload_value: list/dict representing the response of the query
-        :success: boolean
-        :receiverId: string
-
-        '''
-        response_msg = self.get_header(msg_type, recipients=[])
-        response_msg['payload'] = dict()
-        response_msg['payload'][payload_key] = payload_value
-        response_msg['payload']['success'] = success
-        response_msg['payload']['receiverId'] = receiverId
-        return response_msg
-
-
